requirements.txt
selenium==4.18.1
beautifulsoup4==4.12.3
pandas==2.2.1
spacy==3.7.4
nltk==3.8.1
python-dotenv==1.0.1
requests==2.31.0
PyPDF2==3.0.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.3 